# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AgentRun": {
        "additionalProperties": false,
        "properties": {
          "conversationId": {
            "anyOf": [
              {
                "maxLength": 96,
                "minLength": 3,
                "pattern": "^[a-z][a-z0-9_:-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Conversationid"
          },
          "currentNode": {
            "minLength": 1,
            "title": "Currentnode",
            "type": "string"
          },
          "errorCode": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errorcode"
          },
          "finishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Finishedat"
          },
          "id": {
            "maxLength": 96,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_:-]*$",
            "title": "Id",
            "type": "string"
          },
          "inputHash": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Inputhash",
            "type": "string"
          },
          "modelName": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Modelname"
          },
          "organizationId": {
            "maxLength": 96,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_:-]*$",
            "title": "Organizationid",
            "type": "string"
          },
          "promptVersion": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Promptversion"
          },
          "retryCount": {
            "minimum": 0,
            "title": "Retrycount",
            "type": "integer"
          },
          "sourceSnapshotIds": {
            "items": {
              "maxLength": 96,
              "minLength": 3,
              "pattern": "^[a-z][a-z0-9_:-]*$",
              "type": "string"
            },
            "title": "Sourcesnapshotids",
            "type": "array"
          },
          "startedAt": {
            "format": "date-time",
            "title": "Startedat",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          }
        },
        "required": [
          "id",
          "organizationId",
          "currentNode",
          "status",
          "inputHash",
          "sourceSnapshotIds",
          "startedAt",
          "retryCount"
        ],
        "title": "AgentRun",
        "type": "object"
      },
      "AgentRunStep": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maxLength": 96,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_:-]*$",
            "title": "Id",
            "type": "string"
          },
          "node": {
            "minLength": 1,
            "title": "Node",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Payload",
            "type": "object"
          },
          "runId": {
            "maxLength": 96,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_:-]*$",
            "title": "Runid",
            "type": "string"
          },
          "status": {
            "enum": [
              "processing",
              "completed",
              "failed",
              "skipped"
            ],
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "runId",
          "node",
          "status",
          "timestamp",
          "payload"
        ],
        "title": "AgentRunStep",
        "type": "object"
      },
      "AgentRunStepsResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AgentRunStep"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/DataProvenance"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "AgentRunStepsResponse",
        "type": "object"
      },
      "AnalysisRequest": {
        "additionalProperties": false,
        "properties": {
          "assetIds": {
            "items": {
              "maxLength": 96,
              "minLength": 3,
              "pattern": "^[a-z][a-z0-9_:-]*$",
              "type": "string"
            },
            "minItems": 1,
            "title": "Assetids",
            "type": "array"
          },
          "eventId": {
            "maxLength": 96,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_:-]*$",
            "title": "Eventid",
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "assetIds"
        ],
        "title": "AnalysisRequest",
        "type": "object"
      },
      "AnalysisResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AgentRun"
          },
          "meta": {
            "$ref": "#/components/schemas/DataProvenance"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "AnalysisResponse",
        "type": "object"
      },
      "AnalysisStatus": {
        "enum": [
          "processing",
          "completed",
          "insufficient_evidence",
          "failed"
        ],
        "title": "AnalysisStatus",
        "type": "string"
      },
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "anyOf": [
              {
                "maxLength": 96,
                "minLength": 3,
                "pattern": "^[a-z][a-z0-9_:-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Requestid"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "DataMode": {
        "enum": [
          "fixture",
          "live",
          "fallback"
        ],
        "title": "DataMode",
        "type": "string"
      },
      "DataProvenance": {
        "additionalProperties": false,
        "properties": {
          "dataAsOf": {
            "format": "date-time",
            "title": "Dataasof",
            "type": "string"
          },
          "dataMode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "provider": {
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          },
          "retrievedAt": {
            "format": "date-time",
            "title": "Retrievedat",
            "type": "string"
          },
          "warnings": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "dataMode",
          "provider",
          "retrievedAt",
          "dataAsOf",
          "freshness",
          "warnings"
        ],
        "title": "DataProvenance",
        "type": "object"
      },
...
      "Freshness": {
        "additionalProperties": false,
        "properties": {
          "evaluatedAt": {
            "format": "date-time",
            "title": "Evaluatedat",
            "type": "string"
          },
          "isStale": {
            "title": "Isstale",
            "type": "boolean"
          },
          "staleAfterSeconds": {
            "exclusiveMinimum": 0,
            "title": "Staleafterseconds",
            "type": "integer"
          }
        },
        "required": [
          "evaluatedAt",
          "staleAfterSeconds",
          "isStale"
        ],
        "title": "Freshness",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "SseEvent": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maxLength": 96,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_:-]*$",
            "title": "Id",
            "type": "string"
          },
          "node": {
            "minLength": 1,
            "title": "Node",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Payload",
            "type": "object"
          },
          "runId": {
            "maxLength": 96,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_:-]*$",
            "title": "Runid",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "runId",
          "node",
          "status",
          "timestamp",
          "payload"
        ],
        "title": "SseEvent",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Contrato del MVP; la implementación del runtime comienza en Fase 1.",
    "title": "NexoMercado AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyses": {
      "post": {
        "operationId": "createAnalysis",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Analysis accepted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request could not be completed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request could not be completed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request could not be completed"
          }
        },
        "summary": "Start an idempotent market analysis"
      }
    },
    "/api/v1/analyses/{runId}": {
      "get": {
        "operationId": "getAnalysis",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "minLength": 3,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request could not be completed"
          }
        },
        "summary": "Read an analysis run"
      }
    },
    "/api/v1/analyses/{runId}/stream": {
      "get": {
        "operationId": "streamAnalysis",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "minLength": 3,
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "x-eventSchema": {
                  "$ref": "#/components/schemas/SseEvent"
                }
              }
            },
            "description": "Server-Sent Events stream with heartbeat comments"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request could not be completed"
          }
        },
        "summary": "Stream analysis progress with replay support"
      }
    },
...
    "/api/v1/runs/{runId}/steps": {
      "get": {
        "operationId": "listAgentRunSteps",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "minLength": 3,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunStepsResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request could not be completed"
          }
        },
        "summary": "List auditable run steps"
      }
    },
...
  }
}